api: move options to object.go
